Push the processDocumentRecord and processNonStaticDocuments back into the Dems Adapter, and have the reports processor call it.
